Little refactor to separate app_registry from auth
